Update CourseList.java
Change a few method names

Switched to a semester object rather than an int to represent a semester.

I'm not sure about changing semester to an object, but it seems to me that we should have semesters identified by year and Fall/Spring/MayX/... and if they're identified by both of these things, then it's easier to make an object than to code them as integers. We might instead make semesters implicit in the CourseTime class, but I like the idea of letting Schedule, CourseList, and Course see semesters as one type of entity. A fourth option is to make semesters relative to a base year (say the student's freshman year) but that might complicate time comparisons. I'm open to arguments for any of these options, as well as other options.