Simplify requirement
Adds some support for credit hour requirements, which makes the problem of fastestCompletionSet NPHard as it's now a 0-1 knapsack problem where utility = credit hours and weight = number of courses that a subrequirement needs to be complete. Because of that, we simply removed some functionality and it trimmed down a bunch of stuff.