Speed and debug
Fixed some speed issues with terminal requirement's getCreditHours (it was going through the entire course list to find the credit hours every time a terminal wanted to see if it satisfied another requirement. I changed terminals to store the correct value after they've asked courseList once.)         Also fixed a number of bugs in checking for isCompletedBy, isSatisfiedBy, and so on in both Requirement and Terminal Requirement.